requirements.txt
Deprecated==1.2.15
img2pdf==0.5.1
lxml==5.3.0
packaging==24.2
pikepdf==9.4.2
pillow==11.0.0
wrapt==1.17.0
